refactor: Вынесла отдельно хук useCardStatistic
